Working to configure ethernet  🔌
Transitioned from a website that would be hosted on the microcontroller. While the setup for this was easy with the well documented and more used Pico W, it was much more difficult using the webhat pico ethernet adapted. We switched to using the Wizhat w5500 with pico to send POST requests to a server that's planned to be run on the raspberry pi.